// src/models/Employee.model.ts
export interface Employee {
	uid?: string;
	employeeId?: string | null;
	employeeName?: string;
	email?: string;
	employmentStatus?: string;
	teamName?: string;
	teamUid?: string;
	roleName?: string;
	roleUid?: string;
	phoneNo?: string;
	orgKey?: string;
	joiningDate?: string;
	isDeleted?: boolean;
}